优化代理器改为对象提交
