+++ spark-yun-agent/src/main/java/com/isxcode/star/agent/run/impl/StandaloneAgentService.java
@@ -1,9 +1,9 @@
-package com.isxcode.star.agent.run;
+package com.isxcode.star.agent.run.impl;
 
 import com.alibaba.fastjson2.JSON;
+import com.isxcode.star.agent.run.AgentService;
 import com.isxcode.star.api.agent.constants.AgentType;
-import com.isxcode.star.api.agent.pojos.req.DeployContainerReq;
-import com.isxcode.star.api.agent.pojos.req.YagExecuteWorkReq;
+import com.isxcode.star.api.agent.pojos.req.SubmitWorkReq;
 import com.isxcode.star.api.work.constants.WorkType;
 import com.isxcode.star.backend.api.base.exceptions.IsxAppException;
 import lombok.extern.slf4j.Slf4j;
@@ -36,6 +36,11 @@ import java.util.regex.Pattern;
 @Slf4j
 @Service
 public class StandaloneAgentService implements AgentService {
+
+    @Override
+    public String getAgentType() {
+        return AgentType.StandAlone;
+    }
 
     @Override
     public String getMaster(String sparkHomePath) {
@@ -87,56 +92,47 @@ public class StandaloneAgentService implements AgentService {
     }
 
     @Override
-    public SparkLauncher genSparkLauncher(YagExecuteWorkReq yagExecuteWorkReq) {
+    public SparkLauncher getSparkLauncher(SubmitWorkReq submitWorkReq) {
 
-        String appName = "zhiqingyun-" + yagExecuteWorkReq.getWorkType() + "-" + yagExecuteWorkReq.getWorkId() + "-"
-            + yagExecuteWorkReq.getWorkInstanceId();
+        SparkLauncher sparkLauncher =
+            new SparkLauncher().setVerbose(false).setMainClass(submitWorkReq.getSparkSubmit().getMainClass())
+                .setDeployMode("cluster").setMaster(getMaster(submitWorkReq.getSparkHomePath()))
+                .setSparkHome(submitWorkReq.getAgentHomePath() + File.separator + "spark-min");
 
-        SparkLauncher sparkLauncher;
-        if (WorkType.SPARK_JAR.equals(yagExecuteWorkReq.getWorkType())) {
-            // 如果是自定义作业
-            sparkLauncher = new SparkLauncher().setVerbose(false)
-                .setMainClass(yagExecuteWorkReq.getSparkSubmit().getMainClass()).setDeployMode("cluster")
-                .setAppName(yagExecuteWorkReq.getSparkSubmit().getAppName() + "-" + yagExecuteWorkReq.getWorkType()
-                    + "-" + yagExecuteWorkReq.getWorkId() + "-" + yagExecuteWorkReq.getWorkInstanceId())
-                .setMaster(getMaster(yagExecuteWorkReq.getSparkHomePath()))
-                .setAppResource(yagExecuteWorkReq.getAgentHomePath() + File.separator + "file" + File.separator
-                    + yagExecuteWorkReq.getSparkSubmit().getAppResource())
-                .setSparkHome(yagExecuteWorkReq.getAgentHomePath() + File.separator + "spark-min");
+        if (WorkType.SPARK_JAR.equals(submitWorkReq.getWorkType())) {
+            sparkLauncher
+                .setAppName(submitWorkReq.getSparkSubmit().getAppName() + "-" + submitWorkReq.getWorkType() + "-"
+                    + submitWorkReq.getWorkId() + "-" + submitWorkReq.getWorkInstanceId())
+                .setAppResource(submitWorkReq.getAgentHomePath() + File.separator + "file" + File.separator
+                    + submitWorkReq.getSparkSubmit().getAppResource());
         } else {
-            sparkLauncher = new SparkLauncher().setVerbose(false)
-                .setMainClass(yagExecuteWorkReq.getSparkSubmit().getMainClass()).setDeployMode("cluster")
-                .setAppName(appName).setMaster(getMaster(yagExecuteWorkReq.getSparkHomePath()))
-                .setAppResource(yagExecuteWorkReq.getAgentHomePath() + File.separator + "plugins" + File.separator
-                    + yagExecuteWorkReq.getSparkSubmit().getAppResource())
-                .setSparkHome(yagExecuteWorkReq.getAgentHomePath() + File.separator + "spark-min");
+            sparkLauncher
+                .setAppName("zhiqingyun-" + submitWorkReq.getWorkType() + "-" + submitWorkReq.getWorkId() + "-"
+                    + submitWorkReq.getWorkInstanceId())
+                .setMaster(getMaster(submitWorkReq.getSparkHomePath())).setAppResource(submitWorkReq.getAgentHomePath()
+                    + File.separator + "plugins" + File.separator + submitWorkReq.getSparkSubmit().getAppResource());
         }
 
         // 添加额外依赖
-        if (yagExecuteWorkReq.getLibConfig() != null) {
-            yagExecuteWorkReq.getLibConfig().forEach(e -> {
-                String libPath =
-                    yagExecuteWorkReq.getAgentHomePath() + File.separator + "file" + File.separator + e + ".jar";
-                sparkLauncher.addJar(libPath);
-            });
+        if (submitWorkReq.getLibConfig() != null) {
+            submitWorkReq.getLibConfig().forEach(e -> sparkLauncher
+                .addJar(submitWorkReq.getAgentHomePath() + File.separator + "file" + File.separator + e + ".jar"));
+        }
+
+        // 添加Excel/csv文件
+        if (submitWorkReq.getPluginReq().getCsvFilePath() != null) {
+            sparkLauncher.addFile(submitWorkReq.getPluginReq().getCsvFilePath());
         }
 
         // 添加自定义函数
-        if (yagExecuteWorkReq.getPluginReq().getCsvFilePath() != null) {
-            sparkLauncher.addFile(yagExecuteWorkReq.getPluginReq().getCsvFilePath());
+        if (submitWorkReq.getFuncConfig() != null) {
+            submitWorkReq.getFuncConfig().forEach(e -> sparkLauncher.addJar(
+                submitWorkReq.getAgentHomePath() + File.separator + "file" + File.separator + e.getFileId() + ".jar"));
         }
 
-        // 添加自定义函数
-        if (yagExecuteWorkReq.getFuncConfig() != null) {
-            yagExecuteWorkReq.getFuncConfig().forEach(e -> {
-                String libPath = yagExecuteWorkReq.getAgentHomePath() + File.separator + "file" + File.separator
-                    + e.getFileId() + ".jar";
-                sparkLauncher.addJar(libPath);
-            });
-        }
-
-        if (!Strings.isEmpty(yagExecuteWorkReq.getAgentHomePath())) {
-            File[] jarFiles = new File(yagExecuteWorkReq.getAgentHomePath() + File.separator + "lib").listFiles();
+        // 添加至轻云的jar
+        if (!Strings.isEmpty(submitWorkReq.getAgentHomePath())) {
+            File[] jarFiles = new File(submitWorkReq.getAgentHomePath() + File.separator + "lib").listFiles();
             if (jarFiles != null) {
                 for (File jar : jarFiles) {
                     try {
@@ -151,21 +147,23 @@ public class StandaloneAgentService implements AgentService {
             }
         }
 
-        if (WorkType.SPARK_JAR.equals(yagExecuteWorkReq.getWorkType())) {
-            sparkLauncher.addAppArgs(yagExecuteWorkReq.getArgs());
+        // 压缩请求的args
+        if (WorkType.SPARK_JAR.equals(submitWorkReq.getWorkType())) {
+            sparkLauncher.addAppArgs(submitWorkReq.getArgs());
         } else {
             sparkLauncher.addAppArgs(Base64.getEncoder()
-                .encodeToString(yagExecuteWorkReq.getPluginReq() == null ? yagExecuteWorkReq.getArgsStr().getBytes()
-                    : JSON.toJSONString(yagExecuteWorkReq.getPluginReq()).getBytes()));
+                .encodeToString(submitWorkReq.getPluginReq() == null ? submitWorkReq.getArgsStr().getBytes()
+                    : JSON.toJSONString(submitWorkReq.getPluginReq()).getBytes()));
         }
 
-        yagExecuteWorkReq.getSparkSubmit().getConf().forEach(sparkLauncher::setConf);
+        // 将提交的spark配置加入到sparkLauncher
+        submitWorkReq.getSparkSubmit().getConf().forEach(sparkLauncher::setConf);
 
         return sparkLauncher;
     }
 
     @Override
-    public String executeWork(SparkLauncher sparkLauncher) throws IOException {
+    public String submitWork(SparkLauncher sparkLauncher) throws Exception {
 
         Process launch = sparkLauncher.launch();
         InputStream inputStream = launch.getErrorStream();
@@ -200,7 +198,7 @@ public class StandaloneAgentService implements AgentService {
     }
 
     @Override
-    public String getAppStatus(String submissionId, String sparkHomePath) throws IOException {
+    public String getWorkStatus(String submissionId, String sparkHomePath) throws Exception {
 
         Document doc = Jsoup.connect(getMasterWebUrl(sparkHomePath)).get();
 
@@ -224,7 +222,7 @@ public class StandaloneAgentService implements AgentService {
     }
 
     @Override
-    public String getAppLog(String submissionId, String sparkHomePath) throws IOException {
+    public String getStderrLog(String submissionId, String sparkHomePath) throws Exception {
 
         Document doc = Jsoup.connect(getMasterWebUrl(sparkHomePath)).get();
 
@@ -258,7 +256,7 @@ public class StandaloneAgentService implements AgentService {
     }
 
     @Override
-    public String getStdoutLog(String submissionId, String sparkHomePath) throws IOException {
+    public String getStdoutLog(String submissionId, String sparkHomePath) throws Exception {
 
         Document doc = Jsoup.connect(getMasterWebUrl(sparkHomePath)).get();
 
@@ -292,7 +290,7 @@ public class StandaloneAgentService implements AgentService {
     }
 
     @Override
-    public String getAppData(String submissionId, String sparkHomePath) throws IOException {
+    public String getWorkDataStr(String submissionId, String sparkHomePath) throws Exception {
 
         Document doc = Jsoup.connect(getMasterWebUrl(sparkHomePath)).get();
 
@@ -327,7 +325,7 @@ public class StandaloneAgentService implements AgentService {
     }
 
     @Override
-    public void killApp(String submissionId, String sparkHomePath, String agentHomePath) throws IOException {
+    public void stopWork(String submissionId, String sparkHomePath, String agentHomePath) throws Exception {
 
         if (Strings.isEmpty(sparkHomePath) || "null".equals(sparkHomePath)) {
             sparkHomePath = agentHomePath + "/zhiqingyun-agent/spark-min";
@@ -352,48 +350,5 @@ public class StandaloneAgentService implements AgentService {
             log.error(e.getMessage(), e);
             throw new IsxAppException("中止失败");
         }
-    }
-
-    @Override
-    public String getAgentName() {
-        return AgentType.StandAlone;
-    }
-
-    @Override
-    public SparkLauncher genSparkLauncher(DeployContainerReq deployContainerReq) throws IOException {
-
-        String appName = "zhiqingyun-SPARK_CONTAINER-" + deployContainerReq.getContainerId();
-
-        SparkLauncher sparkLauncher = new SparkLauncher().setVerbose(false)
-            .setMainClass(deployContainerReq.getSparkSubmit().getMainClass()).setDeployMode("cluster")
-            .setAppName(appName).setMaster(getMaster(deployContainerReq.getSparkHomePath()))
-            .setAppResource(deployContainerReq.getAgentHomePath() + File.separator + "plugins" + File.separator
-                + deployContainerReq.getSparkSubmit().getAppResource())
-            .setSparkHome(deployContainerReq.getAgentHomePath() + File.separator + "spark-min");
-
-        if (!Strings.isEmpty(deployContainerReq.getAgentHomePath())) {
-            File[] jarFiles = new File(deployContainerReq.getAgentHomePath() + File.separator + "lib").listFiles();
-            if (jarFiles != null) {
-                for (File jar : jarFiles) {
-                    try {
-                        // 使用本地hive驱动
-                        if (!jar.getName().contains("hive") && !jar.getName().contains("zhiqingyun-agent.jar")) {
-                            sparkLauncher.addJar(jar.toURI().toURL().toString());
-                        }
-                    } catch (MalformedURLException e) {
-                        log.error(e.getMessage(), e);
-                        throw new IsxAppException("50010", "添加lib中文件异常", e.getMessage());
-                    }
-                }
-            }
-        }
-
-        sparkLauncher.addAppArgs(Base64.getEncoder()
-            .encodeToString(deployContainerReq.getPluginReq() == null ? deployContainerReq.getArgs().getBytes()
-                : JSON.toJSONString(deployContainerReq.getPluginReq()).getBytes()));
-
-        deployContainerReq.getSparkSubmit().getConf().forEach(sparkLauncher::setConf);
-
-        return sparkLauncher;
     }
 }
