+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/req/GetWorkDataReq.java
@@ -1,0 +1,13 @@
+package com.isxcode.star.api.agent.pojos.req;
+
+import lombok.Data;
+
+@Data
+public class GetWorkDataReq {
+
+    private String appId;
+
+    private String clusterType;
+
+    private String sparkHomePath;
+}
