+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/req/GetWorkStderrLogReq.java
@@ -1,0 +1,13 @@
+package com.isxcode.star.api.agent.pojos.req;
+
+import lombok.Data;
+
+@Data
+public class GetWorkStderrLogReq {
+
+    private String appId;
+
+    private String clusterType;
+
+    private String sparkHomePath;
+}
