+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/req/GetWorkStdoutLogReq.java
@@ -1,0 +1,13 @@
+package com.isxcode.star.api.agent.pojos.req;
+
+import lombok.Data;
+
+@Data
+public class GetWorkStdoutLogReq {
+
+    private String appId;
+
+    private String clusterType;
+
+    private String sparkHomePath;
+}
