+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/req/StopWorkReq.java
@@ -1,0 +1,15 @@
+package com.isxcode.star.api.agent.pojos.req;
+
+import lombok.Data;
+
+@Data
+public class StopWorkReq {
+
+    private String appId;
+
+    private String clusterType;
+
+    private String sparkHomePath;
+
+    private String agentHomePath;
+}
