+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/req/SubmitWorkReq.java
@@ -6,7 +6,7 @@ import lombok.Data;
 import java.util.List;
 
 @Data
-public class YagExecuteWorkReq {
+public class SubmitWorkReq {
 
     private PluginReq pluginReq;
