+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/res/ExecuteContainerSqlRes.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ContainerGetDataRes {
+public class ExecuteContainerSqlRes {
 
     private List<List<String>> data;
