+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/agent/pojos/res/SubmitWorkRes.java
@@ -1,7 +1,6 @@
 package com.isxcode.star.api.agent.pojos.res;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,9 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class YagGetDataRes {
+public class SubmitWorkRes {
 
-    private List<List> data;
-
-    private String applicationId;
+    private String appId;
 }
