+++ spark-yun-backend/spark-yun-api/src/main/java/com/isxcode/star/api/work/constants/WorkType.java
@@ -55,4 +55,9 @@ public interface WorkType {
      * Excel导入作业.
      */
     String EXCEL_SYNC_JDBC = "EXCEL_SYNC_JDBC";
+
+    /**
+     * 实时计算.
+     */
+    String REAL_WORK = "REAL_WORK";
 }
