+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/star/modules/work/run/impl/SparkContainerSqlExecutor.java
@@ -2,7 +2,7 @@ package com.isxcode.star.modules.work.run.impl;
 
 import com.alibaba.fastjson.JSON;
 import com.isxcode.star.api.agent.pojos.req.ExecuteContainerSqlReq;
-import com.isxcode.star.api.agent.pojos.res.ContainerGetDataRes;
+import com.isxcode.star.api.agent.pojos.res.ExecuteContainerSqlRes;
 import com.isxcode.star.api.cluster.constants.ClusterNodeStatus;
 import com.isxcode.star.api.container.constants.ContainerStatus;
 import com.isxcode.star.api.work.constants.WorkLog;
@@ -139,11 +139,11 @@ public class SparkContainerSqlExecutor extends WorkExecutor {
             ExecuteContainerSqlReq executeContainerSqlReq = ExecuteContainerSqlReq.builder()
                 .port(String.valueOf(containerEntityOptional.get().getPort())).sql(script).build();
 
-            ContainerGetDataRes containerGetDataRes;
+            ExecuteContainerSqlRes containerGetDataRes;
             try {
                 containerGetDataRes = new RestTemplate().postForEntity(
                     genHttpUrl(engineNode.getHost(), engineNode.getAgentPort(), "/yag/executeContainerSql"),
-                    executeContainerSqlReq, ContainerGetDataRes.class).getBody();
+                    executeContainerSqlReq, ExecuteContainerSqlRes.class).getBody();
             } catch (Exception e) {
                 log.error(e.getMessage());
                 throw new WorkRunException(e.getMessage());
