+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/star/modules/work/run/impl/SyncWorkExecutor.java
@@ -3,8 +3,8 @@ package com.isxcode.star.modules.work.run.impl;
 import com.alibaba.fastjson.JSON;
 import com.isxcode.star.api.agent.pojos.req.PluginReq;
 import com.isxcode.star.api.agent.pojos.req.SparkSubmit;
-import com.isxcode.star.api.agent.pojos.req.YagExecuteWorkReq;
-import com.isxcode.star.api.agent.pojos.res.YagGetLogRes;
+import com.isxcode.star.api.agent.pojos.req.SubmitWorkReq;
+import com.isxcode.star.api.agent.pojos.res.GetWorkStderrLogRes;
 import com.isxcode.star.api.api.constants.PathConstants;
 import com.isxcode.star.api.cluster.constants.ClusterNodeStatus;
 import com.isxcode.star.api.cluster.pojos.dto.ScpFileEngineNodeDto;
@@ -183,7 +183,7 @@ public class SyncWorkExecutor extends WorkExecutor {
 
         // 开始构建作业
         logBuilder.append(LocalDateTime.now()).append(WorkLog.SUCCESS_INFO).append("开始构建作业  \n");
-        YagExecuteWorkReq executeReq = new YagExecuteWorkReq();
+        SubmitWorkReq executeReq = new SubmitWorkReq();
         executeReq.setWorkId(workRunContext.getWorkId());
         executeReq.setWorkType(WorkType.DATA_SYNC_JDBC);
         executeReq.setWorkInstanceId(workInstance.getId());
@@ -390,8 +390,8 @@ public class SyncWorkExecutor extends WorkExecutor {
                 }
 
                 // 解析日志并保存
-                YagGetLogRes yagGetLogRes =
-                    JSON.parseObject(JSON.toJSONString(baseResponse.getData()), YagGetLogRes.class);
+                GetWorkStderrLogRes yagGetLogRes =
+                    JSON.parseObject(JSON.toJSONString(baseResponse.getData()), GetWorkStderrLogRes.class);
                 logBuilder.append(LocalDateTime.now()).append(WorkLog.SUCCESS_INFO).append("日志保存成功 \n");
                 if (yagGetLogRes != null) {
                     workInstance.setYarnLog(yagGetLogRes.getLog());
